Absorb conductor into analyse
